🧪 Test contract types
